requirements.txt
decouple==0.0.7
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
python-dotenv==1.0.0
tomli==2.0.1
